Allow wider selection for Summoning spell
Increased the possible level range of monsters for the Summoning spell
(there was a problem with the High Priest only summoning Leng Matriarchs
since they were the only eligible monsters).

Also fixed a bug for the "Shapeshifts" property, that prevented
shapeshifting into any monster that had a max dungeon level of -1 (no
limit). A few new monsters should now be possible to shapeshift into,
that was not possible before - e.g. Apes and Leng Spiders.