Simplify weapon stats, add idle animation support
Removed the concept of "plus" damage modifiers for weapons, instead they
now simply have a damage range and a hit chance modifier (the damage
range is still randomized for weapons found on the map, so there are
still "higher quality" versions of weapons).

Weapons are now named like this:
  A Machete (3-11 dmg +0% hit)

Instead of:
  A Machete 7.0 +0%

Also, very loosely related to the change above (since it affected screen
space requirements), the ability for displaying idle animations while
the game is waiting for input has been added. Currently this is used for
"animating" burning tiles or tiles corrupted by Strange Colors, and also
for auto-scrolling item descriptions in the (rare) case that they do not
fit in the window.