Add dlvl limitations for monster group sizes
Limited the group sizes on the first few floors for Reanimated Corpses,
Bloated Corpses, and Wolves.

The most significant change is that Reanimated Corpses always spawn
alone on the first floor (unless multiple groups happen to spawn near
each other), on the second floor there are at most "a few", and
thereafter there may also be a "pack" or "swarm" as usual.