Update to SDL_mixer 2.0.4
This fixes an issue were Mix_SetPanning was not working on Windows,
causing all sound effects to be played at max level regardless of
the master volume option. This problem also meant that there were no
stereo effects in the game (the player is supposed to HEAR from the
audio played that a sound is coming from the east for example, in
addition to the information from the text and graphics).