Keep sounds loaded also at 0% volume
Previously, setting the volume to 0% would deallocate all audio from
memory (it was used as a way to completely disable the audio). Raising
the volume from 0% to 10% required loading the audio files from disk
into memory again (which is very slow).

Now the audio is kept in memory regardless of the volume control
setting, making it much more smooth to change this setting.