Fix hearing sounds from unseen monsters
Previously the origin of a sound was always considered to be seen by the
player if the origin position was seen. This meant that for example a
sneaking monster moving through water did not yield any message (since
this message is skipped if the origin is seen, to avoid message spam),
although audio would still be played.

With this change, if a creature is the "originator" of a sound, then the
origin of the sound is considered to be seen if the player can see this
creature - otherwise if the sound was not made by a creature, the origin
is seen if the origin position is seen.