Fix/unify spell reflection implementation
Some spells (such as the knockback spell cast by monsters) could bypass
spell shield of the original caster when a spell was reflected back on
them. This is because those spells handled spell reflection differently
- they simply swapped caster and target then proceeded to apply the
spell effect(s).

Now all spells re-casts the effect again with swapped target and caster
(which is how most spells already worked), so that the new effect
procedure considers spell shield on the new target.