Fix double game time ticking bug
When the player kicked monsters with "w", game time was accidentally
ticked twice for this action (in different functions), which meant that
both the player's and some other creature's turn was ended (whichever
happened to come after the player in the action order).

Also removed game time ticking when player takes an insanity hit, which
did not make sense, and might perhaps have lead to strange behavior
since it was not called at a proper moment.

In debug mode, there is now also a mechanism that asserts that tick is
not called multiple times during the same creature's act phase.