Fix and simplify wall tile processing
Summary: Remembered wall tiles outside the player's FOV could suddenly
change if terrain was destroyed or created (even if unseen).

Wall tiles are set to a "front" or "top" wall tile depending on what is
in the position below the wall. For example a wall with a floor in the
position below should be drawn as a "front" wall (the image showing the
front of a brick wall).

This was previously handled during map drawing as a "post processing
step". However this handling was broken, since it was reading actual
terrain objects on the map even for unseen positions - meaning that a
wall tile in an unseen remembered position could suddenly change if
other terrain was destroyed or created.

Now the wall terrain objects handle this processing themselves, which
allowed significantly simpler code, with the above error fixed.