Fix Prolonged Life bugs
The Exorcist trait Prolonged Life was completely broken - it could give
SP to the player, or kill the player even though they actually had SP
available.

This was cauesd by two coding errors: 1) an inverted condition, and 2)
not catching the return value of the function handling prolonged life
which adjusted the damage taken.