Prioritize bashing terrain before creatures
Previously if there was a creature in a position, this was prioritized
for kicking over kicking/bashing the terrain in that position. This was
a problem since an unknown invisible creature could occupy a door for
example, which would cause the player to unexpectedly and
unintentionally kick the creature instead of the door.

A consequence of this change is that ethereal creatures inside blocking
terrain cannot be kicked.