Fix problem with monsters delayed by liquid
Attacking a monster that is delayed by moving into liquid no longer
allows them to act immediately.

The problem was that attacking a monster cleared the "waiting" status
effect (since directly attacking a monster should cause it to react
immediately).

Handled the situation by adding a new status effect "delayed by liquid",
which is applied instead of "waiting". The "waiting" effect is now
intended to be used only for AI control (like forcing a summoned monster
to wait for a turn before acting).