Fix creature speed problem in game time system
Fixed a problem with the game time system, that caused slow creatures to
act slightly more often and fast/very fast creatures to act slightly
less often than they were supposed to (especially noticeable when kiting
a slow monster, since it would occasionally get unexpected extra moves).

The issue was that the actors delays were decremented at the wrong place
in the loop that handles actor timing. The result of this was that for
example a normal speed creature acted once every 7 delay ticks (should
be 6) and a fast creature acted once every 4 delay ticks (should be
3). In other words a fast creature acted 7 times for every 4 actions of
a normal speed creatures. (And creatures of other speeds had equivalent
problems.)