Show when status effects are ending
Status effects that will end immediately after the next action are shown
as e.g. "Premonition (ending)" in the gui, so that it is possible to
know whether it will last until the next player turn or not (previously
seeing an active status effect listed could be misleading since it could
end before any monster acts).