Fix weapon unload bug
When unloading a weapon from the ground, the unloaded ammo was cast to
an item::Ammo pointer. In the case of unloading spike guns, this is
undefined behavior, since spikes are not instances of item::Ammo, but of
item::Wpn.

Fixed by not casting to item::Ammo, and instead keeping the unloaded
ammo as just a pointer to item::Item (there was no reason for the cast).