Revert to background colors for multiple objects
When there are multiple objects in the same position, such as an item on
top of a trap, the game now uses background colors again to signify
this (instead of blinking items as in 581f569e).

Also made some improvements to how the game represents multiple things
in the same position, mainly that a different background color is used
when there are multiple corpses, or when a corpse is hidden under an
item.