Fix drawing of remembered positions
Drawing of remembered tiles to the bottom and right of the map panel was
too limited. Other objects (such as seen terrain or monsters) can be
drawn as a fraction of a tile, but for remembered tiles (terrain or
items), this was not possible.

When drawing remembered positions, the tiles are now also drawn a little
bit outside the viewport to allow showing a fraction of tiles if the map
panel size is not aligned with a whole number of map tiles (for example
15.6 map tiles can be shown on the Y axis). The drawing is clipped to
the map panel, so pixels outside the map panel will not be drawn.