Allow attacking through some terrain
Terrain on the path to the target aimed at is considered blocking if it
blocks projectiles, instead of if it's non-open terrain.

Rationale: We do not want to check for "open terrain" (floor-like
terrain), since the player should be able to attack through some
non-open terrain such as barred gate. It is probably a good enough
approximation that you can attack with a long reach weapon through
terrain that you can also shoot through.

However you still cannot attempt to attack a creature inside non-open
terrain that you are unaware of. I.e. you can attack a Reanimated Corpse
on the other side of a closed barred gate, but you cannot attack an
unknown Ghost on the gate.