Fix algorithms for things randomly blocking the map
The spell side effect that created trees/fungi, and the Warping
Aberrance effect that alters the environment had several bugs that could
disconnect the map.

* The spell side effect did not consider stairs as free positions,
  meaning they could be surrounded by trees.
* The Warping Aberrance effect DID consider stairs to be free, but this
  is also a problem, because the stairs might then end up blocking other
  positions (since they are not actually passable terrain).
* There were also some similar problems with levers (must be reachable).

Fixed by adding an extra parameter in map_parsing::is_map_connected, to
specify extra positions that must be reachable, but are not necessarily
free themselves (stairs, levers).