Make some monsters wait on seeing the player
A few monsters wait for a turn when they see the player and are aware of
the player. This only happens once, until the monster loses awareness
again. It is intended to avoid some unfair instant deaths by allowing
the player a chance to act first (for example Cultists and White Spiders
have this behavior).