Fix long action interruption bug
Fixed a bug that caused interruptions of longer player actions (handling
armor or using the medical bag).

The problem was that status effects that interrupt the player (fainting,
burning, entangled, stuck, etc) would interrupt regardless of which
creature it was applied to.

The effect of this was probably most often noticeable when monsters
outside view got stuck in mud.