Give the moribund status a duration
The Flagellant "moribund" status (low health bonuses) is now applied for
a number of turns when taking damage that reduces health to 6 hit points
or below (previously it was always active while hit points were below
the limit, and ended immediately when restored above the limit).

The reason for this change is to allow the player to utilize the effect
for a bit even if hit points are immediately restored above the limit
(e.g. when using the Shed Impurity spell).