Fix viewing known spells when picking trait
Viewing information about known spells while picking a new trait was
handled poorly if the player did not know any spells.

A message would be printed that the player did not know any spells, and
then the game waited for input in the trait picking menu, but was still
drawing the message.

The problem was that states ending immediately in the "on_start" hook
was not considered by the state handler.

Now the message is printed with a "more" prompt, and then you go back to
the trait menu.