Reserve keys in the "apply" menu
In the "apply" menu, "a" is always reserved for the Electric Lantern and
'b' for the Medical Bag (and 's' for the Holy Symbol if Exorcist). On
the first screen of the "apply" menu, these items will always appear on
top and their keys will never be used for any other item.