Update hit punctuation calculation
The hit punctuation ('.', '!' or '!!!') is now calculated relative to a
fixed value (the damage required to cause a wound), instead of relative
to the weapon max damage.

Damage required to wound is 5.

Damage from 1 to 4 is "Minor damage" ('.')
Damage from 5 to 7 is "Medium damage" ('!')
Damage from 8 and above is "Major" damage ('!!!')