Fix weapon damage percentage scaling
When scaling weapon damage due to things like the Weak status effect,
the range and "plus" values were scaled and rounded individually. This
could lead to for example a "50%" damage reduction actually reducing the
damage a lot more than that.

For example "3-9 +1" (avg 6.5) damage became "1-4 +0" (avg 2.5) when
scaled by "50%".

The fix was to scale the final rolled value instead.