Fix burrowing effect
The burrowing status effect (now available via a potion) would allow the
player to move into a wall on the final turn of the effect, and then it
ended without destroying that wall - leaving the player on top of a
wall.

Now burrowing also destroys the current terrain on the turn that it
ends.