Make Azathoth's Gaze always run sound with message
Previously a sound containing a message was only triggered if the player
was casting the spell and there were no seen targets. This made it hard
to interpret what was going on.

Now such a sound is always triggered regardless of who is casting and if
any targets are seen.