Allow specifying relative required levels for group sizes
The required dungeon level for monster group sizes in monsters.xml can
now also be specified relative to the minimum dungeon level of the
monster (with format "+N" levels).

For example if a monster has min_dungeon_level "4", then a group_size
entry with required_dungeon_level="+2" will cause the monster to start
spawning with that group size on dungeon level 6.